requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
python-dateutil==2.9.0.post0
pandas==2.2.2
openpyxl==3.1.5
gunicorn==22.0.0
psycopg2-binary==2.9.9
